refactor(project): apply new file structure
- move page components to pages/ dir
- move simple components to components/ dir
- move styled components to styles files
- move routes to separate file
- rename cont text404 to TEXT404
