+++ src/components/BoldText/index.js
@@ -1,0 +1,10 @@
+import {styled} from "@mui/material/styles";
+
+
+export const BoldText = styled('span')(
+    ({ theme }) => `
+        font-weight: ${theme.typography.fontWeightBold};
+    `,
+);
+
+export default BoldText;
