+++ src/pages/PaymentAndDelivery/index.js
@@ -1,8 +1,7 @@
 import Box from "@mui/material/Box";
-import Divider from '@mui/material/Divider';
 import React from "react";
 import Typography from "@mui/material/Typography";
-import {styled} from '@mui/material/styles';
+import {DividerStyled} from "./styles";
 
 
 const Foreword = () => (
@@ -48,15 +47,6 @@ const PaymentText = () => (
             Оплата производится после согласования цены, обсуждения всех деталей заказа и до поступления его в обработку.
         </Typography>
     </Box>
-);
-
-const DividerStyled = styled(Divider)(
-    ({ theme }) => `
-        width: 107%;
-        margin-top: 3rem;
-        background-color: ${theme.palette.primary.main};
-        align-self: center;
-    `,
 );
 
 const DeliveryText = () => (
